fix: unblock CSLM CI on Scala 2.12 and JDK21 APIs
Motivation:
PR #4886 failed in CI due to two blockers: Scala 2.12 type/reassignment errors in CSLM spliterators and JEP431 SequencedMap tests compiling on non-JDK21 source sets.

Modification:
- Made CSLM spliterator comparator helper Scala-2.12-friendly by avoiding fragile generic inference in compare calls.
- Renamed spliterator constructor forwarding params to avoid shadowing inherited mutable state fields used by trySplit.
- Moved SequencedMap-specific tests out of shared suites into require-jdk21 test source and left shared-suite pointers.

Result:
- javalib2_12/compile succeeds again.
- testsJVM2_12/test:compile and testsJVM3/test:compile succeed with SequencedMap tests still covered under JDK21.
