+++ unit-tests/shared/src/test/scala/org/scalanative/testsuite/javalib/util/concurrent/ConcurrentSkipListSubMapTest.scala
@@ -113,69 +113,8 @@ class ConcurrentSkipListSubMapTest extends JSR166Test {
     mustEqual(java.lang.Long.MAX_VALUE, entries.estimateSize())
   }
 
-  @Test def testSubMapSequencedMapViews(): Unit = {
-    val map = map5()
-
-    mustEqual(iOne, map.sequencedKeySet().getFirst())
-    mustEqual(iFive, map.sequencedKeySet().reversed().getFirst())
-
-    mustEqual("A", map.sequencedValues().getFirst())
-    mustEqual("E", map.sequencedValues().reversed().getFirst())
-
-    mustEqual(iOne, map.sequencedEntrySet().getFirst().getKey())
-    mustEqual("A", map.sequencedEntrySet().getFirst().getValue())
-    mustEqual(iFive, map.sequencedEntrySet().reversed().getFirst().getKey())
-    mustEqual("E", map.sequencedEntrySet().reversed().getFirst().getValue())
-
-    mustEqual(iFive, map.reversed().firstEntry().getKey())
-    mustEqual(iOne, map.reversed().reversed().firstEntry().getKey())
-  }
-
-  @Test def testSubMapSequencedMapPutFirstLastUnsupported(): Unit = {
-    val map = map5()
-    assertThrows(
-      classOf[UnsupportedOperationException],
-      map.putFirst(null, "X")
-    )
-    assertThrows(
-      classOf[UnsupportedOperationException],
-      map.putFirst(iTwo, null)
-    )
-    assertThrows(
-      classOf[UnsupportedOperationException],
-      map.putFirst(iTwo, "X")
-    )
-    assertThrows(
-      classOf[UnsupportedOperationException],
-      map.putLast(iThree, "X")
-    )
-  }
-
-  @Test def testSubMapSequencedValuesRemoveFirstLast(): Unit = {
-    val map = map5()
-
-    mustEqual("A", map.sequencedValues().removeFirst())
-    assertFalse(map.containsKey(iOne))
-    mustEqual(iTwo, map.firstKey())
-
-    mustEqual("E", map.sequencedValues().removeLast())
-    assertFalse(map.containsKey(iFive))
-    mustEqual(iFour, map.lastKey())
-  }
-
-  @Test def testSubMapSequencedEntrySetRemoveFirstLast(): Unit = {
-    val map = map5()
-
-    val first = map.sequencedEntrySet().removeFirst()
-    mustEqual(iOne, first.getKey())
-    mustEqual("A", first.getValue())
-    assertFalse(map.containsKey(iOne))
-
-    val last = map.sequencedEntrySet().removeLast()
-    mustEqual(iFive, last.getKey())
-    mustEqual("E", last.getValue())
-    assertFalse(map.containsKey(iFive))
-  }
+  // SequencedMap API tests are in require-jdk21:
+  // javalib/util/concurrent/JEP431_ConcurrentSkipListMapTestOnJDK21.scala
 
   @Test def testSubMapValuesRemoveIfRemovesConditionally(): Unit = {
     val map = map5()
